Fix unstable mv test cases
